projects/amazon-seller/src/requirements.txt
# Amazon Seller AI - Dependencies

# Amazon SP-API
python-amazon-sp-api>=1.0.0
boto3>=1.28.0
requests>=2.31.0

# Web Framework (for OAuth server)
flask>=3.0.0

# Data Processing
pandas>=2.0.0

# AI Services
anthropic>=0.18.0

# Environment
python-dotenv>=1.0.0

# Shared utilities (installed from ../shared/)
# - grok_image_gen (XAI)
# - gmail_monitor (Google)
# - revenue_analytics (Google Sheets)
# - twilio_sms (Twilio)
